pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tensorpotential"
version = "0.5.2"  #### UPD in setup.py to avoid confusion
authors = [
    { name = "Anton Bochkarev", email = "anton.bochkarev@rub.de" },
    { name = "Yury Lysogorskiy", email = "yury.lysogorskiy@rub.de" },
]
description = "Graph Atomic Cluster Expansion (GRACE)"
readme = "README.md"
requires-python = ">=3.9,<3.13" # Update if needed
license = { text = "Academic Software License (ASL)" }
classifiers = [
    "License :: OSI Approved :: Academic Free License (AFL)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "tensorflow[and-cuda]<2.20",
    "tf_keras",  # Required for legacy Keras API. Users may need to set TF_USE_LEGACY_KERAS=1
    "scipy",
    "matscipy",
    "numpy<3.0.0",
    "sympy",
    "pandas<3.0.0",
    "ase",
    "pyyaml>=6.0.2",
    "tqdm",
]
[project.urls]
"Homepage" = "https://github.com/ICAMS/grace-tensorpotential"

#[project.scripts]
#gracemaker = "bin:gracemaker"
#grace_models = "bin:grace_models"
#grace_collect = "bin:grace_collect"
#extxyz2df = "bin:extxyz2df"
#df2extxyz = "bin:df2extxyz"
#grace_preprocess = "bin:grace_preprocess"
